requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
audiosegment==0.23.0
bcrypt==4.1.3
blinker==1.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.6
filelock==3.13.4
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-text-splitters==0.2.0
langsmith==0.1.59
llvmlite==0.42.0
MarkupSafe==2.1.5
marshmallow==3.21.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
orjson==3.10.3
packaging==23.2
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
SQLAlchemy==2.0.30
sympy==1.12
tenacity==8.3.0
tiktoken==0.6.0
torch==2.3.0
tqdm==4.66.2
triton==2.3.0
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
webrtcvad==2.0.10
Werkzeug==3.0.2
yarl==1.9.4
